+++ CHANGELOG.md
@@ -211,3 +211,4 @@
 * Define `Instruct` pattern
 * Define `Load` pattern
 * Change `Functor` typeclass definition
+* Replace `Primary` associated type family with `Casted` open type family